resolves issue #88
