+++ apps/frontend/components/client/ChannelCard.tsx
@@ -1,15 +1,6 @@
 import Image from 'next/image';
 import { Body, BodySmall, Flex, cn } from '@river/design-system';
-import { Hex } from 'viem';
-
-export interface Channel {
-  name: string;
-  description?: string;
-  creator: Hex;
-  members?: string[];
-  cover: string;
-}
-
+import { type Channel } from '../../types/types';
 
 interface ChannelCardProps extends React.HTMLAttributes<HTMLDivElement> {
   channel: Channel;
